Adds driving functionality to Car class
Extends the Car class to include a Drive method. This method utilizes the AboutCar method to output the car's brand and model, along with a message indicating that the car is now driving.

Additionally, creates a Program class to demonstrate the usage of the Car class and its new Drive method by instantiating Car objects and calling their Drive methods.